Reformat tables
Replace unusual whitespace characters and realign tables
when cleaning a source document.

Also emit warnings when two tables are right next to each other,
as this may mean it is the same table just separated by a page break.

(Marker's output, or the PDF source, includes tabs and <br> in tables,
presumably to try to preserve the original look of the table?)