Realign tables automatically
In order to do this, we have to normalise whitespace to spaces (and not
the frankly baffling presence of "<br>") first. Possible table run-on
is flagged (a table across a page breaks, but marker has rendered it
as a new table).

Also cleans some unusual unicode choices (eg "...spell–" in table header
was sometimes coming out as "combining long stroke overlay", so it
looked like ł). Other unusual unicode is flagged.