Edit files based on warnings
A script that, provided with errors including line numbers, will edit
the SRD file being checked at that line, so it can be manually
edited if necessary.

A file of messages to ignore can also be provided, so when a possible
error is not an error, you don't keep editing the same spot.

Used like:

    python clean_srd.py --warn ...srd \
        | ./edit_warnings ...srd.md ...ignore.txt