Warn for unusual bullets and no em-dash spacing
Most bullets are converted to hyphens for Markdown by `marker`, but
it misses some. Check for those characters.

Thin/hair spaces around em-dashes aren't a thing in Markdown, so we
can't use that to space them out, but the visual separation is needed
in my opinion so we'll put up with wider space.