Breakdown now ignores changes and header levels
When breaking down a file that's already been broken down and had
header levels fixed in the fragments, running breakdown again would
tediously change the files even when "nothing" had changed.

It now compares files and only writes files that have changed, and
the `-i` option tells it to ignore header levels when comparing.
Additionally, if a file has changed, it will attempt to fix the
header levels in the new fragment.