Update the breakdown script
I don't know why I had the parts to fix and check headers separate,
but now it does that as part of the breakdown.

And in SRD 5.2.1, there is a need to append some part to an existing
file (class and subclass being separated by spell lists) so that has
been modified too.